refactor: split monolithic sidepanel.js into focused modules
- Extract helpers.js: escHtml, fullName, fmtMoney, fmtDate, formatPhone, icon, el
- Extract q10Api.js: sendMsg (chrome.runtime.sendMessage wrapper)
- Extract state.js: mutable state (currentPhone, currentResult, wizardState, catalogsCache)
  plus tag/note CRUD (getContactTags, setContactTags, getContactNotes, addContactNote, deleteContactNote)
- Extract renderer.js: all render functions (renderEstudiante, renderContacto,
  renderOportunidad, renderUnknown, renderLoading, renderError, renderNoConversation,
  renderTagsSection, renderNotesSection, attachTagsAndNotes, generateCSV, downloadCSV)
- Extract handlers.js: event binding (showToast, bindExportButtons, bindTagsHandlers,
  bindNotesHandlers, bindManualSearch, bindRetryHandler, searchPhone, refreshFinancials,
  showGenerarCobroModal, showCreateOportunidadModal, showCreateContactoModal,
  showCreateActividadModal, bindEstudianteActions, bindOportunidadActions,
  bindContactoActions, bindUnknownActions, startEnrollmentWizard)
- Extract wizard.js: full 5-step enrollment wizard (renderWizardStep,
  submitWizardStep, renderWizardComplete, cancelWizard)
- sidepanel.js: thin orchestration layer (~120 lines) — imports modules
  and wires up chrome.storage.onChanged listener and init()

Fixes diogenesmendes01/q10-whatsapp-plugin#8
